refactor sources; some phrma updates
