+++ frontend/src/cards/ui/SourcesInfo.tsx
@@ -1,0 +1,42 @@
+import React from 'react'
+import { DATA_CATALOG_PAGE_LINK } from '../../utils/internalRoutes'
+import {
+  LinkWithStickyParams,
+  DATA_SOURCE_PRE_FILTERS,
+} from '../../utils/urlutils'
+import styles from './Sources.module.scss'
+import { type DataSourceInfo, insertPunctuation } from './SourcesHelpers'
+
+interface SourcesInfoProps {
+  dataSourceMap: Record<string, DataSourceInfo>
+}
+
+export default function SourcesInfo(props: SourcesInfoProps) {
+  return (
+    <p className={styles.FootnoteText}>
+      Sources:{' '}
+      {Object.keys(props.dataSourceMap).map((dataSourceId, idx) => (
+        <React.Fragment key={dataSourceId}>
+          <LinkWithStickyParams
+            target="_blank"
+            to={`${DATA_CATALOG_PAGE_LINK}?${DATA_SOURCE_PRE_FILTERS}=${dataSourceId}`}
+          >
+            {props.dataSourceMap[dataSourceId].name}
+          </LinkWithStickyParams>{' '}
+          {props.dataSourceMap[dataSourceId].updateTimes.size === 0 ? (
+            <>(last update unknown) </>
+          ) : (
+            <>
+              (updated{' '}
+              {Array.from(props.dataSourceMap[dataSourceId].updateTimes).join(
+                ', '
+              )}
+              )
+            </>
+          )}
+          {insertPunctuation(idx, Object.keys(props.dataSourceMap).length)}
+        </React.Fragment>
+      ))}{' '}
+    </p>
+  )
+}
