+++ frontend/src/pages/WhatIsHealthEquity/FaqTab.tsx
@@ -4,7 +4,7 @@ import { Typography } from '@mui/material'
 import { CONTACT_TAB_LINK } from '../../utils/internalRoutes'
 import { Helmet } from 'react-helmet-async'
 import MoreHorizIcon from '@mui/icons-material/MoreHoriz'
-import { CITATION_APA } from '../../cards/ui/Sources'
+import { CITATION_APA } from '../../cards/ui/SourcesHelpers'
 import { selectFAQs } from './FaqData'
 
 function FaqTab() {
